pyproject.toml
[tool.poetry]
name = "pycoinlab"
version = "0.1.0"
description = "Projeto de estudos para implementação de uma blockchain em Python"
authors = ["Bruno Almeida <bruno.a.s.almeida@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
flask-migrate = "^4.0.7"
flask-restx = "^1.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-flask = "^1.3.0"
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
